Fix "block build-specific files" regex.
It wasn't anchored in any way, so for example the pattern '.libs' was
matching 'discover.libsnapclient.actions'.

It now checks that '.libs' and others are either at the beginning of the
path or preceded by a slash, so it should match .libs/x or x/.libs/y
but not foo.libs.