Phabricator tool: normalize URIs visibility
Summary:
Hide all the extra URIs added by Phabricator (and not really used)
and set as visible only the URI mirrored from anongit.
It will probably be useless, or at least it will need some changes,
when the repositories will be handled by Phabricator directly.

Test Plan:
Run the script with -a where at least two repositories have URIs with the
wrong visibility and ensure that they are fixed.
Run again, but when the URIs have the correct visibility and ensure that
nothing is changed.
Repeat the test above without -a, but specifying directly the ID of
the repositories with -r and ensure that only those repositories are
affected or at least processed.

Reviewers: #sysadmin, lbeltrame

Reviewed By: #sysadmin, lbeltrame

Subscribers: bcooksley, lbeltrame, #sysadmin

Differential Revision: https://phabricator.kde.org/D3633